refactored ResultEntry and ResultList by separating out the CategoryRenderingMaps
