+++ client/src/components/ResultEntry/styles.tsx
@@ -1,4 +1,4 @@
-import styled, { css } from "styled-components";
+import styled from "styled-components";
 
 const ResultOuterWrapper = styled.div`
   display: flex;
@@ -44,6 +44,72 @@ const TagIcon = styled.img.attrs((props) => ({
   cursor: pointer;
 `;
 
+const DetailsWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  margin-left: 5px;
+`;
+
+const Title = styled.div`
+  font-size: 16px;
+  font-weight: bolder;
+  &:hover {
+    text-decoration: underline;
+    cursor: pointer;
+  }
+`;
+
+const Date = styled.div`
+  font-size: 12px;
+  color: #585858;
+  margin-right: 20px;
+`;
+
+const Details = styled.div`
+  font-size: 14px;
+  font-weight: lighter;
+`;
+
+const IconWrapper = styled.div`
+  display: flex;
+  align-items: center;
+`;
+
+const ContactIcon = styled.img.attrs({
+  src: "./icons/contact-icon.png",
+})`
+  width: 20px;
+  height: 20px;
+`;
+
+const CalendarIcon = styled.img.attrs({
+  src: "./icons/Calendar-icon.png",
+})`
+  width: 20px;
+  height: 20px;
+`;
+
+const DropboxIcon = styled.img.attrs({
+  src: "./icons/Dropbox-icon.png",
+})`
+  width: 20px;
+  height: 20px;
+`;
+
+const SlackIcon = styled.img.attrs({
+  src: "./icons/Slack-icon.png",
+})`
+  width: 25px;
+  height: 25px;
+`;
+
+const TwitterIcon = styled.img.attrs({
+  src: "./icons/twitter-icon.png",
+})`
+  width: 20px;
+  height: 18px;
+`;
+
 export {
   ResultOuterWrapper,
   ResultLeftWrapper,
@@ -52,4 +118,14 @@ export {
   PinIcon,
   TagWrapper,
   TagIcon,
+  DetailsWrapper,
+  Title,
+  Date,
+  Details,
+  IconWrapper,
+  ContactIcon,
+  CalendarIcon,
+  DropboxIcon,
+  SlackIcon,
+  TwitterIcon,
 };
